refactor: split types/consensus into a module tree; move LightClientStore to the engine
types/consensus.rs was a 540-line grab-bag. Split it into three
concern-focused submodules behind a thin re-export root, so every
existing `crate::types::consensus::X` path keeps resolving:

- headers.rs   — BeaconBlockHeader, fork-aware LightClientHeader + variants,
                 Capella/Deneb execution payload headers.
- committee.rs — SyncCommittee (+ committee-root views), SyncAggregate.
- messages.rs  — LightClientUpdate / FinalityUpdate / SyncCommitteeUpdate,
                 LightClientBootstrap.

LightClientStore moves out of the wire-type module to
consensus/store.rs: it is the engine's verified processing state, not a
data type exchanged on the wire. processor.rs imports it from its new
home; each type's unit tests travel with it.

No public API change (re-exports preserve all paths); no logic change.
