+++ src/consensus/mod.rs
@@ -3,6 +3,7 @@
 pub mod bls;
 pub mod merkle;
 pub mod processor;
+pub(crate) mod store;
 pub mod sync_committee;
 
 #[cfg(test)]
